localscore/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "localscore"
version = "1.0.0"
description = "CLI benchmarking tool for measuring LLM inference performance"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Mozilla Foundation"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "llama-cpp-python>=0.2.0",
    "psutil>=5.9.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
nvidia = ["pynvml>=11.0.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
localscore = "localscore.cli:main"

[project.urls]
Homepage = "https://github.com/Mozilla-Ocho/llamafile"
Repository = "https://github.com/Mozilla-Ocho/llamafile"

[tool.setuptools.packages.find]
where = ["."]
